refactor: board 구조 개선
