+++ src/test/java/chess/domain/piece/PieceDirectionTest.java
@@ -1,7 +1,6 @@
 package chess.domain.piece;
 
-import chess.domain.piece.PieceDirection;
-import chess.domain.piece.exception.WrongDirectionException;
+import chess.domain.exception.WrongDirectionException;
 import chess.domain.square.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
